More pages

+++ app/api/request/[id]/route.ts
@@ -1,12 +1,40 @@
 import { PrismaClient } from "@prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 const prisma = new PrismaClient();
 
+export async function GET(request: NextRequest, { params }: { params: { id: string } }): Promise<NextResponse> {
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.email) {
+        return NextResponse.json({}, { status: 401 });
+    }
+
+    const r = await prisma.paymentRequest.findUnique({
+        where: {
+            id: params.id,
+            owner: {
+                email: session.user.email,
+            },
+        },
+    });
+
+    return NextResponse.json({ request: r }, { status: !r ? 404 : 200 });
+}
+
 export async function DELETE(request: NextRequest, { params }: { params: { id: string } }): Promise<NextResponse> {
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.email) {
+        return NextResponse.json({}, { status: 401 });
+    }
+
     await prisma.paymentRequest.delete({
         where: {
             id: params.id,
+            owner: {
+                email: session.user.email,
+            },
         },
     });
 
@@ -14,7 +42,10 @@ export async function DELETE(request: NextRequest, { params }: { params: { id: s
 }
 
 export async function PATCH(request: NextRequest, { params }: { params: { id: string } }): Promise<NextResponse> {
-    console.log("id", params);
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.email) {
+        return NextResponse.json({}, { status: 401 });
+    }
 
     const body = (await request.json()) as {
         name: string;
@@ -27,6 +58,9 @@ export async function PATCH(request: NextRequest, { params }: { params: { id: st
     const newRequest = await prisma.paymentRequest.update({
         where: {
             id: params.id,
+            owner: {
+                email: session.user.email,
+            },
         },
         data: {
             description: body.description,
@@ -34,6 +68,5 @@ export async function PATCH(request: NextRequest, { params }: { params: { id: st
         },
     });
 
-    // prisma.user.delete()
     return NextResponse.json({ request: newRequest });
 }
